+++ src/PSTT.Dashboard.Client/Components/SidePanelEnums.cs
@@ -1,0 +1,16 @@
+namespace PSTT.Dashboard.Components;
+
+public enum SidePanelTab
+{
+    NodeProps,
+    AddNode,
+    DataExplorer,
+    PageProps,
+    DashboardProps
+}
+
+public enum SubView
+{
+    List,
+    Custom
+}